% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipart.match.plot.R
\name{bipart.match.plot}
\alias{bipart.match.plot}
\title{Plot bipartite matches}
\usage{
bipart.match.plot(runid, lasset, obset, plotdir = NULL)
}
\arguments{
\item{runid}{list. Names of sampling areas cross-referenced in lasset and obset}

\item{lasset}{list. List of point cloud datasets in LAS format to apply bipartite matching to. Typically the output of an optimized ITC segementation or delineation algorithm, such as Li2012.opt.}

\item{obset}{dataframe. Dataset containing at minimum X,Y,Z coordinates for field-identified trees with an additional column labeling trees by an ID matching one of those in runid.}
}
\value{
dataframe. Performance and accuracy statistics for each sample area, parameter permutation, and algorithm.
}
\description{
Run optimal bipartite matching on field and modeled trees by minimizing Euclidean distances. Return data for plotting.
}
