% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bipart.match3.R
\name{bipart.match3}
\alias{bipart.match3}
\title{Bipartite matching}
\usage{
bipart.match3(runid, lasset, obset, return.matches = F, plotdir = F)
}
\arguments{
\item{runid}{list. Names of sampling areas cross-referenced in lasset and obset}

\item{lasset}{list. List of point cloud datasets in LAS format to apply bipartite matching to. Typically the output of an optimized ITC segementation or delineation algorithm, such as Li2012.opt.}

\item{obset}{dataframe. Dataset containing at minimum X,Y,Z coordinates for field-identified trees with an additional column labeling trees by an ID matching one of those in runid.}

\item{return.matches}{logical. Specify whether to return raw matching data or summary statistics. If T, a dataframe with raw matching data is returned for downstream analysis. If F, summary statistics are returned. Default is F.}

\item{plotdir}{character. Destination directory path for figures produced when computing summary statistics. Default is F. Only relevant if return.matches is F.}
}
\value{
dataframe. Performance and accuracy statistics for each sample area, parameter permutation, and algorithm.
}
\description{
Run optimal bipartite matching on field and modeled trees by minimizing Euclidean distances. Return performance statistics and data for plotting.
}
