% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li2012_funs.R
\name{li2012.init}
\alias{li2012.init}
\title{Li 2012 initialization}
\usage{
li2012.init(pc, dt1, dt2, R, Zu, hmin)
}
\arguments{
\item{pc}{LAS. Point cloud data in LAS format}

\item{dt1}{numeric vector. Set of values for threshold number 1. See reference page 79 in Li et al. (2012). Default is 1.5.}

\item{dt2}{numeric vector. Set of values for threshold number 2. See reference page 79 in Li et al. (2012). Default is 2.}

\item{R}{numeric vector. Set of values for search radius. See page 79 in Li et al. (2012). Default is 2. If R = 0 all the points are automatically considered as local maxima and the search step is skipped (much faster).}

\item{Zu}{numeric. If point elevation is greater than Zu, dt2 is used, otherwise dt1 is used. See page 79 in Li et al. (2012). Default is 15.}

\item{hmin}{numeric. Minimum height of a detected tree. Default is 2.}
}
\value{
dataframe. X,Y,Z values corresponding to locations and heights of detected trees
}
\description{
Initialize Li 2012 for optimization
}
