% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/li2012_funs.R
\name{li2012.opt}
\alias{li2012.opt}
\title{Li 2012 optimization}
\usage{
li2012.opt(x, params, hmin = 1.3)
}
\arguments{
\item{x}{list. List of point cloud datsets in LAS format}
}
\value{
list. X,Y,Z locations of modeled trees in each sample area
}
\description{
Run Li 2012 optimization algorithm across sample areas
}
