% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lmfvw_funs.R
\name{lmf.vw.init}
\alias{lmf.vw.init}
\title{LMF variable window function}
\usage{
lmf.vw.init(pc, b0, b1, c0, shape, hmin)
}
\description{
Define variable window as thresholded exponential f(height)

Initialize LMF fixed window for optimization
}
