% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.models.R
\name{make.modframe}
\alias{make.modframe}
\title{make.modframe}
\usage{
make.modframe(y, df, modtype = c("gam", "gbm"), target.vars, itx = c("none"))
}
\description{
Make a standard model frame for inferential modeling
}
