% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multichm_funs.R
\name{mc.init}
\alias{mc.init}
\title{LayerStacking initialization}
\usage{
mc.init(pc, res, layer_thickness, dist_2d, dist_3d, use_max = F, ws = 5)
}
\description{
Initialize LayerStacking for optimization
}
