% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gams.R
\name{pe.pivot.itx}
\alias{pe.pivot.itx}
\title{Pivot sliced interaction model frame}
\usage{
pe.pivot.itx(sdf, plot.vars)
}
\description{
Pivot an interaction partial-effects dataframe longer for comparison of
multiple partial effects
}
