% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.gams.R
\name{pe.slice.itx}
\alias{pe.slice.itx}
\title{Slice model frame for interactions}
\usage{
pe.slice.itx(v, mod)
}
\description{
Slice a model frame for prediction of an
interaction's partial effects over the interacting variables'
observed intervals
}
