% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrees_funs.R
\name{ptrees.init}
\alias{ptrees.init}
\title{Ptrees initialization}
\usage{
ptrees.init(k, pc, hmin, nmax = 9L)
}
\arguments{
\item{k}{integer vector. A serie of k-nearest neighbors to use. In this original paper a k refers to a 'scale' of analyse (see reference).}

\item{pc}{LAS. Point cloud data in LAS format}

\item{hmin}{scalar. This is an addition from the original paper to limit oversegmentation. Point below this threshold cannot initiate new trees or increase a hull (see details). Set to -Inf to strictly respect original paper.}

\item{nmax}{integer. This is an addition from the original paper to protect against uncomputable cases (see details). Set to +Inf to strictly respect the original paper (not recommended)}
}
\value{
dataframe. X,Y,Z values corresponding to locations and heights of detected trees
}
\description{
Initialize PTrees for optimization
}
