% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptrees_funs.R
\name{ptrees.opt}
\alias{ptrees.opt}
\title{PTrees optimization}
\usage{
ptrees.opt(x, params, hmin = 1.3)
}
\arguments{
\item{x}{list. List of point cloud datsets in LAS format}

\item{params}{list. List of optimizable values for k in ptrees}
}
\value{
list. X,Y,Z locations of modeled trees in each sample area
}
\description{
Run PTrees optimization algorithm across sample areas
}
